requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-decouple==3.8
openpyxl==3.1.2
